Lua executor fixes
Previous release had to change the behavior of how script.Script was
parsed. The change related on how the content of the script was saved.
It didn't keep headers which caused issues at execution. These changes
makes it we don't keep just the content but we pass around the
script.Script struct around. This change should help with the execution
speed a little as we don't have to parse the script on every execution.

All stores now keep the parsed script internally. For most cases it
doesn't affect the user but for the etcd store it will require to delete
everything currently in etcd before reapplying with an updated version
of msgscript-cli.