Big refactor, fixes both wasm and podman
Now the scripts are fetched when the message is received and each scripts
gets called with their respective executor. This simplifies greatly the
executors as they no longer have to fetch the scripts themselves. They
retain the store if they need to fetch more things (like libraries for
the lua executor).